Expose minimum supported rust-version via pub function
The original intention for this was to be used by `forc` to include a
compatible `rust-version` field in the manifest of the generated test
harnesses. However, I think specifying the minimum `rust-version` in
`fuels` itself is enough, and will still correctly notify the user as
`fuels` is a dependency of the generated test harnesses anyway.